:recycle: Cleaning interfaces / Renaming things
