+++ src/constants/trading.constants.ts
