+++ src/errors/trading.errors.ts
