+++ src/interfaces/trading.interfaces.ts
