+++ src/routes/trade.routes.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from 'express';
 import { readAccount } from '../services/account.service';
 import { Trade } from '../entities/trade.entities';
-import { getTradeSide } from '../utils/trade.utils';
+import { getTradeSide } from '../utils/trading.utils';
 import { HttpCode } from '../constants/http.constants';
-import { TRADE_EXECUTION_SUCCESS } from '../messages/trade.messages';
-import { TradingService } from '../services/trade.service';
+import { TRADE_EXECUTION_SUCCESS } from '../messages/trading.messages';
+import { TradingService } from '../services/trading/trading.service';
 
 export const postTrade = async (req: Request, res: Response): Promise<void> => {
   const { direction, stub, symbol }: Trade = req.body;
